Set region decoding to disabled by default and enable it for FullWidthImageItem()
