+++ android/ui/post/src/main/java/any/ui/post/item/FullWidthImageItem.kt
@@ -32,6 +32,7 @@ import androidx.compose.ui.unit.sp
 import any.base.image.ImageRequest
 import any.base.image.ImageState
 import any.ui.common.image.AsyncImage
+import any.ui.common.image.RegionDecodeThresholds
 import any.ui.common.theme.darkerImagePlaceholder
 import any.ui.common.theme.imagePlaceholder
 
@@ -111,6 +112,7 @@ internal fun FullWidthImageItem(
             },
             reloadKey = reloadFactor,
             showProgressbar = true,
+            regionDecodeThresholds = RegionDecodeThresholds.Default,
         )
 
         if (!isLoaded && showIndex) {
